requirements.txt
Django~=3.0.4
Pillow~=7.0.0